perf(graph): linear recorded-trace lowering for large captures
Lowering a real 79,306-record Dynamo capture was stuck in prepare for
>17 min; the whole file now decodes and lowers in ~22s (byte-parity
preserved), matching the Python adapter.

Two hot-path fixes:

- BlockHash is now `i128` instead of `num_bigint::BigInt`. Recorded
  cache-block hashes are u64 plus small negative virtual ids, so `i128`
  covers the entire realistic domain as a `Copy` machine integer with an
  identical decimal `Display` (content seeds unchanged). This removes the
  per-hash heap allocation, string-radix parse, clone, compare, and
  HashMap hashing across the millions of block hashes in a large trace.
  `scalar::hash_i128` fast-paths JSON integers and falls back to the
  arbitrary-precision coercion only for the rare oversized/string forms.

- Prompt-message emission now reuses a per-trace
  `(parent, role, block-hashes) -> handle` cache. Shared conversation
  prefixes previously re-decoded (tokenizer) and re-hashed (blake3) the
  full inherited prompt for every turn; caching the interned handle makes
  that the linear splice the Python driver performs. Interning a repeated
  message is a SegmentPool dedup no-op, so the pool output is identical.

All 92 graph unit tests, the WEKA<->Dynamo byte-equivalence parity test,
and the transport tests pass. Verified end-to-end through the Python
aiperf frontend -> aiperf-runner on the full production capture.

+++ crates/aiperf-graph/src/recorded/mod.rs
@@ -24,17 +24,20 @@ use std::error::Error;
 use std::fmt::{self, Display};
 
 use aiperf_dataset::LoadConfig;
-use num_bigint::BigInt;
 
 pub use dynamo::compile_dynamo_trace_input;
 pub use weka::compile_weka_trace_input;
 
 /// Opaque recorded cache-block identity.
 ///
-/// Python accepts arbitrary-size integers. `BigInt` preserves that open domain
-/// without narrowing WEKA hashes to a machine integer; Dynamo's synthetic
-/// negative identities use the same representation.
-pub(crate) type BlockHash = BigInt;
+/// Dynamo/WEKA captures record u64 cache-block hashes, and Dynamo mints small
+/// negative virtual identities for non-replay turns; `i128` losslessly covers
+/// that entire domain as a `Copy` machine integer. This replaces the former
+/// `BigInt` representation, which heap-allocated, parsed, cloned, compared, and
+/// hashed every one of the millions of per-block hashes on the lowering hot
+/// path. `i128::Display` is identical to `BigInt::Display` for every recorded
+/// value, so content-seed derivation stays byte-exact.
+pub(crate) type BlockHash = i128;
 
 /// Corpus used to reconstruct recorded token blocks.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
